# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/prs/{id}/merge:
    post:
      operationId: mergePR
      parameters:
      - description: PR number.
        in: path
        name: id
        required: true
        schema:
          description: PR number.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergePRRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergePRResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Squash-merge a pull request
      tags:
      - prs
  /api/v1/prs/{id}/resolve-comments:
    post:
      operationId: resolveComments
      parameters:
      - description: PR number.
        in: path
        name: id
        required: true
        schema:
          description: PR number.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveCommentsResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Resolve review threads on a pull request
      tags:
      - prs
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    MergePRRequest:
      properties:
        expectedHeadSha:
          pattern: ^[0-9a-fA-F]{40}([0-9a-fA-F]{24})?$
          type: string
        prUrl:
          minLength: 1
          type: string
      required:
      - prUrl
      type: object
    MergePRResponse:
      properties:
        method:
          type: string
        ok:
          type: boolean
        prNumber:
          type: integer
      required:
      - ok
      - prNumber
      - method
      type: object
...
    ResolveCommentsResponse:
      properties:
        ok:
          type: boolean
        resolved:
          type: integer
      required:
      - ok
      - resolved
      type: object
...